backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bycrypt]
python-jose
python-dotenv